Add test for get_file_listing
download_nlm_data.get_file_listing used to be coupled to the FTP
connection object and also had not been tested. It's been
reconfigured to only expect a list of filenames and rewritten
to eliminate bugs in the logic for skipping uninteresting files
as well as the regex used to identify these uninteresting files.