Use FTPFileParams as a namedtuple more often
Previously, FTPFileParams was converted to a dictionary midway through
processing downloads. This wasn't necessary and complicated the
creation of tests for the download_nlm_data module.

Now, FTPFileParams is a namedtuple throughout download_nlm_data, though
it is still converted (using vars() and not namedtuple._asdict()) in
nlm_downloads_db.

This commit also includes relevant changes to the tests for
download_nlm_data and nlm_downloads_db.